+++ src/pages/social/ClubSocialPage.tsx
@@ -2,7 +2,8 @@ import ClubSocialItem from '@/domains/social/ui/ClubSocialItem';
 import SearchTab from '@/domains/search/ui/SearchTab';
 import { useSearchParams } from 'react-router-dom';
 import { useClubSNSByType } from '@/domains/social/api/useClubSNS';
-import { useEffect, useRef } from 'react';
+import { useScrollRestoration } from '@/hooks/useScrollRestoration';
+import { useInfiniteScroll } from '@/hooks/useInfiniteScroll';
 import ErrorIcon from '@/assets/common/error-icon.svg?react';
 import {
     CentralCategoryDropdown,
@@ -43,51 +44,16 @@ export default function ClubSocialPage() {
         size: 12,
     });
 
-    const SCROLL_KEY = 'social:scroll';
-    const restoreTarget = useRef<{ y: number; pageCount: number } | null>(null);
-    const hasRestored = useRef(false);
+    const { saveScroll } = useScrollRestoration({
+        storageKey: 'social:scroll',
+        pageCount,
+        isLoading,
+        isFetchingNextPage,
+        hasNextPage,
+        fetchNextPage,
+    });
 
-    useEffect(() => {
-        const saved = sessionStorage.getItem(SCROLL_KEY);
-        if (!saved) return;
-        try {
-            restoreTarget.current = JSON.parse(saved);
-        } catch {
-            // sessionStorage 값이 올바른 JSON 형식이 아닐 경우 무시
-        }
-        sessionStorage.removeItem(SCROLL_KEY);
-    }, []);
-
-    useEffect(() => {
-        const target = restoreTarget.current;
-        if (!target || hasRestored.current || isLoading || isFetchingNextPage)
-            return;
-
-        if (pageCount < target.pageCount && hasNextPage) {
-            fetchNextPage();
-            return;
-        }
-
-        window.scrollTo({ top: target.y, behavior: 'instant' });
-        hasRestored.current = true;
-        restoreTarget.current = null;
-    }, [pageCount, isLoading, isFetchingNextPage, hasNextPage, fetchNextPage]);
-
-    useEffect(() => {
-        const handleScroll = () => {
-            if (
-                window.innerHeight + document.documentElement.scrollTop >=
-                document.documentElement.offsetHeight - 1000
-            ) {
-                if (hasNextPage && !isFetchingNextPage) {
-                    fetchNextPage();
-                }
-            }
-        };
-
-        window.addEventListener('scroll', handleScroll);
-        return () => window.removeEventListener('scroll', handleScroll);
-    }, [hasNextPage, isFetchingNextPage, fetchNextPage]);
+    useInfiniteScroll({ hasNextPage, isFetchingNextPage, fetchNextPage });
 
     const updateSearchParam = (searchKey: string, value: string) => {
         const newParams = new URLSearchParams(searchParams);
@@ -190,13 +156,7 @@ export default function ClubSocialPage() {
                                 clubLogoUrl={account.profileImage}
                                 clubSNSId={account.accountName}
                                 onClick={() => {
-                                    sessionStorage.setItem(
-                                        SCROLL_KEY,
-                                        JSON.stringify({
-                                            y: window.scrollY,
-                                            pageCount,
-                                        }),
-                                    );
+                                    saveScroll();
                                     window.location.href =
                                         account.instagramProfileUrl;
                                 }}
